feat: add authcontroller to gateway
