--- Server/Auth/Sources/App/configure.swift
@@ -1,8 +1,0 @@
-import Fluent
-import Vapor
-
-public func configure(_ app: Application) throws {
-    try configureDatabase(for: app)
-    try configureMigrations(for: app)
-    try configureRoutes(for: app)
-}
