--- Server/Auth/Sources/App/routes.swift
@@ -1,6 +1,0 @@
-import Vapor
-
-func configureRoutes(for app: Application) throws {
-    let authController = AuthController()
-    try app.register(collection: authController)
-}
