--- Server/Auth/Sources/Run/main.swift
@@ -1,11 +1,0 @@
-import App
-import Vapor
-
-var env = try Environment.detect()
-try LoggingSystem.bootstrap(from: &env)
-let app = Application(env)
-
-try configure(app)
-try app.run()
-
-defer { app.shutdown() }
