--- Server/Auth/Tests/AppTests/AppTests.swift
@@ -1,15 +1,0 @@
-@testable import App
-import XCTVapor
-
-final class AppTests: XCTestCase {
-    func testHelloWorld() throws {
-        let app = Application(.testing)
-        defer { app.shutdown() }
-        try configure(app)
-
-        try app.test(.GET, "hello") { res in
-            XCTAssertEqual(res.status, .ok)
-            XCTAssertEqual(res.body.string, "Hello, world!")
-        }
-    }
-}
