+++ Server/Gateway/Sources/App/Controllers/AuthController.swift
@@ -2,7 +2,7 @@
 //  File.swift
 //  
 //
-//  Created by Aidar Nugmanov on 5/5/20.
+//  Created by Aidar Nugmanov on 5/7/20.
 //
 
 import Fluent
@@ -56,3 +56,4 @@ struct AuthController: RouteCollection {
         try request.auth.require(User.self)
     }
 }
+
