+++ Server/Gateway/Sources/App/Models/User.swift
@@ -2,8 +2,10 @@
 //  File.swift
 //  
 //
-//  Created by Aidar Nugmanov on 5/5/20.
+//  Created by Aidar Nugmanov on 5/7/20.
 //
+
+import Foundation
 
 import Fluent
 import Vapor
