+++ Server/Gateway/Sources/App/Models/UserToken.swift
@@ -2,7 +2,7 @@
 //  File.swift
 //  
 //
-//  Created by Aidar Nugmanov on 5/5/20.
+//  Created by Aidar Nugmanov on 5/7/20.
 //
 
 import Fluent
