+++ Server/Gateway/Sources/App/configure.swift
@@ -8,6 +8,7 @@
 import Vapor
 
 public func configure(_ app: Application) throws {
-    try configureMiddleware(for: app)
+    try configureDatabase(for: app)
+    try configureMigrations(for: app)
     try configureRoutes(for: app)
 }
