+++ Server/Gateway/Sources/App/databases.swift
@@ -1,8 +1,8 @@
 //
 //  File.swift
+//  
 //
-//
-//  Created by Temirkhan Sailau on 4/30/20.
+//  Created by Aidar Nugmanov on 5/7/20.
 //
 
 import Fluent
