+++ Server/Gateway/Sources/App/routes.swift
@@ -8,6 +8,9 @@
 import Vapor
 
 func configureRoutes(for app: Application) throws {
+    let authController = AuthController()
+    try app.register(collection: authController)
+    
     let gatewayController = GatewayController()
     try app.register(collection: gatewayController)
 }
