+++ docker-compose.yml
@@ -20,26 +20,12 @@ services:
     env_file:
       - .env
     depends_on:
-      - auth
-      - feedback
-      - content
-      - feed
+      - auth_db
+        # - feedback
+        #- content
+        #- feed
     ports:
       - '8080:80'
-    command: ["serve", "--env", "production", "--hostname", "0.0.0.0", "--port", "80"]
-
-  auth:
-    image: auth:latest
-    build: 
-      context: .
-      args:
-        service_folder: Server/Auth
-    environment:
-      <<: *shared_environment
-    env_file:
-      - .env
-    depends_on:
-      - auth_db 
     command: ["serve", "--env", "production", "--hostname", "0.0.0.0", "--port", "80"]
 
   feed:
